made impl a static inner class. Fixed gradle to make samza-test import samza-operator. Added math commons
