+++ samza-api/src/main/java/org/apache/samza/operators/Windows.java
@@ -225,4 +225,72 @@ public final class Windows {
     return new SessionWindow<>(sessionKeyFunction, (m, c) -> { if (c == null) {return 1; } else { return c+1; }});
   }
 
+
+  static class WindowStateImpl<WV> implements WindowState<WV> {
+
+    private long firstSystemTimeNanos;
+    private long lastSystemTimeNanos;
+    private long earliestEventTimeNanos;
+    private long latestEventTimeNanos;
+    private long numMessages;
+    private WV outputValue;
+
+
+    @Override
+    public long getFirstMessageTimeNs() {
+      return firstSystemTimeNanos;
+    }
+
+    @Override
+    public long getLastMessageTimeNs() {
+      return lastSystemTimeNanos;
+    }
+
+    @Override
+    public long getEarliestEventTimeNs() {
+      return earliestEventTimeNanos;
+    }
+
+    @Override
+    public long getLatestEventTimeNs() {
+      return latestEventTimeNanos;
+    }
+
+    @Override
+    public long getNumberMessages() {
+      return numMessages;
+    }
+
+    @Override
+    public WV getOutputValue() {
+      return outputValue;
+    }
+
+    @Override
+    public void setOutputValue(WV value) {
+      this.outputValue = value;
+    }
+
+    public WindowStateImpl(long firstSystemTimeNanos, long lastSystemTimeNanos, long earliestEventTimeNanos, long latestEventTimeNanos, long numMessages, WV outputValue) {
+      this.firstSystemTimeNanos = firstSystemTimeNanos;
+      this.lastSystemTimeNanos = lastSystemTimeNanos;
+      this.earliestEventTimeNanos = earliestEventTimeNanos;
+      this.latestEventTimeNanos = latestEventTimeNanos;
+      this.numMessages = numMessages;
+      this.outputValue = outputValue;
+    }
+
+    @Override
+    public String toString() {
+      return "WindowStateImpl{" +
+          "firstSystemTimeNanos=" + firstSystemTimeNanos +
+          ", lastSystemTimeNanos=" + lastSystemTimeNanos +
+          ", earliestEventTimeNanos=" + earliestEventTimeNanos +
+          ", latestEventTimeNanos=" + latestEventTimeNanos +
+          ", numMessages=" + numMessages +
+          ", outputValue=" + outputValue +
+          '}';
+    }
+  }
+
 }
